server/pyproject.toml
[tool.poetry]
name = "mcp-server"
version = "0.1.0"
description = "Extensible Multi-Client Platform (MCP) Server with semantic routing and plugin architecture"
authors = ["MCP Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "==4.0.1"
python-multipart = "^0.0.9"
redis = "^5.0.1"
openai = "^1.3.0"
anthropic = "^0.58.0"
langchain = "^0.3.0"
structlog = "^23.2.0"
prometheus-client = "^0.19.0"
python-dotenv = "^1.0.0"
watchdog = "^3.0.0"
selenium = "^4.15.0"
linkedin-scraper = {git = "https://github.com/stickerdaniel/linkedin_scraper.git"}
aiohttp = "^3.9.0"
fastmcp = "^2.10.1"
inquirer = "^3.4.0"
keyring = "^25.6.0"
# PDF Resume Parser dependencies
pdfminer-six = "^20231228"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 